rewrite to TS
